requirements.txt
streamlit==1.39.0
gradio==4.44.0
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.7
langchain-google-genai==2.0.2
chromadb==0.5.11
sentence-transformers==3.2.1
python-dotenv==1.0.1
